+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/dto/TagDto.java
@@ -1,6 +1,7 @@
 package com.itachallenge.challenge.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;