+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/repository/TagRepository.java
@@ -12,6 +12,4 @@ public interface TagRepository extends ReactiveMongoRepository<TagDocument, UUID
 
     Mono<TagDocument> findByTagName(String tagName);
 
-    Mono<Void> deleteByTagName(String tagName);
-
 }