+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/ChallengeServiceImp.java
@@ -268,7 +268,7 @@ public class ChallengeServiceImp implements IChallengeService {
                                         existingLanguage,
                                         savedSolution.getUuid(),
                                         topic,
-                                        tagService.convertStringNameToTag(challengeCreateDto.getTags()));
+                                        challengeCreateDto.getTags());
                                 return challengeRepository.save(challenge)
                                         .map(savedChallenge -> challengeConverter.convertDocumentToDto(challenge,
                                                 ChallengeDto.class));
@@ -276,7 +276,7 @@ public class ChallengeServiceImp implements IChallengeService {
                 });
     }
 
-    private ChallengeDocument buildChallengeDocument(ChallengeCreateDto dto, LanguageDocument language, UUID solutionId, Topic topic, List<TagDocument> tagsAssigned) {
+    private ChallengeDocument buildChallengeDocument(ChallengeCreateDto dto, LanguageDocument language, UUID solutionId, Topic topic, List<UUID> tags) {
         DetailDocument detail = new DetailDocument(dto.getDescription());
 
         return ChallengeDocument.builder()
@@ -287,7 +287,7 @@ public class ChallengeServiceImp implements IChallengeService {
                 .languages(Set.of(language))
                 .solutions(List.of(solutionId))
                 .topic(topic)
-                .tags(tagsAssigned)
+                .tags(tags)
                 .build();
     }
 
@@ -364,9 +364,9 @@ public class ChallengeServiceImp implements IChallengeService {
         Mono<UUID> languageIdMono = validateUUID(String.valueOf(userId));
 
         return Mono.zip(challengeIdMono, languageIdMono)
-                .flatMap(tuple -> {
-                    UUID challengeUuid = tuple.getT1();
-                    UUID userUuid = tuple.getT2();
+                .flatMap(Uuidtuple -> {
+                    UUID challengeUuid = Uuidtuple.getT1();
+                    UUID userUuid = Uuidtuple.getT2();
 
                     return challengeRepository.findByUuid(challengeUuid)
                             .switchIfEmpty(Mono.error(new ChallengeNotFoundReturn404Exception(String.format(CHALLENGE_NOT_FOUND_ERROR, challengeUuid))))
@@ -375,7 +375,8 @@ public class ChallengeServiceImp implements IChallengeService {
                                     .flatMap(isAddedToUsersFavorites -> {
                                         if (Boolean.TRUE.equals(isAddedToUsersFavorites) ||
                                                 Optional.ofNullable(challenge.getTimesFavorite()).orElse(0) == 0) {
-                                            return updateTimesFavorite(challenge);
+                                            challenge.increaseTimesFavorite();
+                                            return challengeRepository.save(challenge);
                                         }
                                         return Mono.just(challenge);
                                     })
@@ -383,11 +384,30 @@ public class ChallengeServiceImp implements IChallengeService {
                 });
     }
 
-    private Mono<ChallengeDocument> updateTimesFavorite(ChallengeDocument challenge) {
-        challenge.setTimesFavorite(
-                Optional.ofNullable(challenge.getTimesFavorite()).orElse(0) + 1
-        );
-        return challengeRepository.save(challenge);
+    @Override
+    public Mono<FavoriteDto> removeChallengeFromFavorites(String challengeId, String userId) {
+        Mono<UUID> challengeIdMono = validateUUID(String.valueOf(challengeId));
+        Mono<UUID> languageIdMono = validateUUID(String.valueOf(userId));
+
+        return Mono.zip(challengeIdMono, languageIdMono)
+                .flatMap(Uuidtuple -> {
+                    UUID challengeUuid = Uuidtuple.getT1();
+                    UUID userUuid = Uuidtuple.getT2();
+
+                    return challengeRepository.findByUuid(challengeUuid)
+                            .switchIfEmpty(Mono.error(new ChallengeNotFoundReturn404Exception(String.format(CHALLENGE_NOT_FOUND_ERROR, challengeUuid))))
+                            .flatMap(challenge -> userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())
+                                    .onErrorResume(throwable -> Mono.error(new InternalServerErrorException(throwable.getMessage())))
+                                    .flatMap(isRemovedFromUsersFavorites -> {
+                                        if (Boolean.TRUE.equals(isRemovedFromUsersFavorites) ||
+                                                Optional.ofNullable(challenge.getTimesFavorite()).orElse(0) == 0) {
+                                            challenge.decreaseTimesFavorite();
+                                            return challengeRepository.save(challenge);
+                                        }
+                                        return Mono.just(challenge);
+                                    })
+                                    .map(savedChallenge -> new FavoriteDto(false, savedChallenge.getTimesFavorite())));
+                });
     }
 
 }