+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/IChallengeService.java
@@ -39,4 +39,6 @@ public interface IChallengeService {
     Mono<ChallengeListDto> getChallengesByTopic(Topic topic, int page, int size);
 
     Mono<FavoriteDto> addChallengeToFavorites(String challengeId, String userId);
+
+    Mono<FavoriteDto> removeChallengeFromFavorites(String challengeId, String userId);
 }