+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/controller/ChallengeControllerTest.java
@@ -1,16 +1,14 @@
 package com.itachallenge.challenge.controller;
 
 import com.itachallenge.challenge.config.PropertiesConfig;
+import com.itachallenge.challenge.document.DetailDocument;
 import com.itachallenge.challenge.dto.*;
 import com.itachallenge.challenge.enums.DifficultyLevel;
 import com.itachallenge.challenge.enums.Topic;
-import com.itachallenge.challenge.exception.ChallengeNotFoundReturn404Exception;
-import com.itachallenge.challenge.exception.InternalServerErrorException;
-import com.itachallenge.challenge.exception.LanguageNotFoundException;
-import com.itachallenge.challenge.exception.ChallengeNotFoundException;
+import com.itachallenge.challenge.exception.*;
 import com.itachallenge.challenge.service.IChallengeService;
+import com.itachallenge.challenge.service.ITagService;
 import com.itachallenge.challenge.service.JwtService;
-import com.itachallenge.challenge.service.TagService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -50,7 +48,7 @@ class ChallengeControllerTest {
     private IChallengeService challengeService;
 
     @MockBean
-    private TagService tagService;
+    private ITagService tagService;
 
     @MockBean
     private DiscoveryClient discoveryClient;
@@ -364,7 +362,7 @@ class ChallengeControllerTest {
 
     @Test
     void addChallenge_test_validRequest() {
-        List<String> tags = List.of("POO");
+        List<UUID> tags = List.of(UUID.randomUUID());
         ChallengeCreateDto formData = new ChallengeCreateDto("títol", "descripció",
                 DifficultyLevel.valueOf("EASY"), "Java", "solució", Topic.LISTS,tags );
 
@@ -385,7 +383,7 @@ class ChallengeControllerTest {
 
     @Test
     void addChallenge_test_emptyField_statusBadRequest() {
-        List<String> tags = List.of("POO");
+        List<UUID> tags = List.of(UUID.randomUUID());
         ChallengeCreateDto formData = new ChallengeCreateDto("",
                 "descripció",
                 DifficultyLevel.valueOf("EASY"),
@@ -405,7 +403,7 @@ class ChallengeControllerTest {
 
     @Test
     void addChallenge_test_invalidLanguage_statusBadRequest() {
-        List<String> tags = List.of("POO");
+        List<UUID> tags = List.of(UUID.randomUUID());
         ChallengeCreateDto formData = new ChallengeCreateDto("títol",
                 "descripció",
                 DifficultyLevel.valueOf("EASY"),
@@ -492,16 +490,16 @@ class ChallengeControllerTest {
     void addChallengeToFavorites_Success_Returns200() {
         String challengeId = "existing_challengeId";
         String userId = "existing_userId";
-        String token = "JWT_token_containing_userId";
+        String authHeader = "validAuthHeader";
 
         FavoriteDto expectedResponse = new FavoriteDto(true, 20);
 
         when(challengeService.addChallengeToFavorites(challengeId, userId)).thenReturn(Mono.just(expectedResponse));
-        when(jwtService.extractUuid(token)).thenReturn(userId);
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenReturn(userId);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
-                .header("Authorization", "Bearer " + token)
+                .header("Authorization", authHeader)
                 .exchange()
                 .expectStatus().isOk()
                 .expectBody(FavoriteDto.class)
@@ -514,16 +512,16 @@ class ChallengeControllerTest {
     void addChallengeToFavorites_ChallengeNotFound_Returns404() {
         String challengeId = "nonExisting_challengeId";
         String userId = "existing_userId";
-        String token = "JWT_token_containing_userId";
+        String authHeader = "validAuthHeader";
 
         String errorMessage = "ErrorMessage";
 
         when(challengeService.addChallengeToFavorites(challengeId, userId)).thenReturn(Mono.error(new ChallengeNotFoundReturn404Exception(errorMessage)));
-        when(jwtService.extractUuid(token)).thenReturn(userId);
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenReturn(userId);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
-                .header("Authorization", "Bearer " + token)
+                .header("Authorization", authHeader)
                 .exchange()
                 .expectStatus().isNotFound()
                 .expectBody(MessageDto.class)
@@ -536,16 +534,16 @@ class ChallengeControllerTest {
     void addChallengeToFavorites_InternalServerError_Returns500() {
         String challengeId = "Existing_challengeId";
         String userId = "existing_userId";
-        String token = "JWT_token_containing_userId";
+        String authHeader = "validAuthHeader";
 
         String errorMessage = "ErrorMessage";
 
         when(challengeService.addChallengeToFavorites(challengeId, userId)).thenReturn(Mono.error(new InternalServerErrorException(errorMessage)));
-        when(jwtService.extractUuid(token)).thenReturn(userId);
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenReturn(userId);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
-                .header("Authorization", "Bearer " + token)
+                .header("Authorization", authHeader)
                 .exchange()
                 .expectStatus()
                 .isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
@@ -558,18 +556,19 @@ class ChallengeControllerTest {
     @Test
     void addChallengeToFavorites_InvalidHeader_Returns400() {
         String challengeId = "Existing_challengeId";
-        String token = "BadToken";
+        String authHeader = "badHeader";
+        String errorMessage = "ErrorMessage";
 
-        when(jwtService.extractUuid(token)).thenReturn(null);
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenThrow(new JwtException(errorMessage));
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
-                .header("Authorization", "Bearer " + token)
+                .header("Authorization", authHeader)
                 .exchange()
                 .expectStatus()
                 .isEqualTo(HttpStatus.BAD_REQUEST)
                 .expectBody(MessageDto.class)
-                .value(messageDto -> Assertions.assertEquals("Invalid Authorization header content", messageDto.getMessage()));
+                .value(messageDto -> Assertions.assertEquals(errorMessage, messageDto.getMessage()));
 
         verify(challengeService, times(0)).addChallengeToFavorites(anyString(), anyString());
     }
@@ -577,6 +576,9 @@ class ChallengeControllerTest {
     @Test
     void addChallengeToFavorites_MissingHeader_Returns400() {
         String challengeId = "Existing_challengeId";
+        String errorMessage = "ErrorMessage";
+
+        when(jwtService.getUserUuIdFromAuthenticationHeader(null)).thenThrow(new JwtException(errorMessage));
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
@@ -584,14 +586,15 @@ class ChallengeControllerTest {
                 .expectStatus()
                 .isEqualTo(HttpStatus.BAD_REQUEST)
                 .expectBody(MessageDto.class)
-                .value(messageDto -> Assertions.assertEquals("Missing or bad formatted Authorization header", messageDto.getMessage()));
+                .value(messageDto -> Assertions.assertEquals(errorMessage, messageDto.getMessage()));
 
         verify(challengeService, times(0)).addChallengeToFavorites(anyString(), anyString());
     }
 
     @Test
+
     @DisplayName("GET recibir respuesta 200 a getTags")
-    void testGetAllTagsEndpoint() {
+    void getTags_test_validRequest() {
 
         TagDto tag1 = new TagDto(UUID.randomUUID(), "POO", "Programación orientada a objetos");
         TagDto tag2 = new TagDto(UUID.randomUUID(), "Algoritmos", "Retos de lógica y eficiencia");
@@ -616,6 +619,141 @@ class ChallengeControllerTest {
                 .jsonPath("$.limit").isEqualTo(2);
 
         verify(tagService).getAllTags();
+    }
+
+    void removeChallengeFromFavorite_Success_Returns200() {
+        String challengeId = "existing_challengeId";
+        String userId = "existing_userId";
+        String authHeader = "validAuthHeader";
+
+        FavoriteDto expectedResponse = new FavoriteDto(false, 20);
+
+        when(challengeService.removeChallengeFromFavorites(challengeId, userId)).thenReturn(Mono.just(expectedResponse));
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenReturn(userId);
+
+        webTestClient.delete()
+                .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
+                .header("Authorization", authHeader)
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody(FavoriteDto.class)
+                .isEqualTo(expectedResponse);
+
+        verify(challengeService, times(1)).removeChallengeFromFavorites(challengeId, userId);
+    }
+
+    @Test
+    void removeChallengeFromFavorite_ChallengeNotFound_Returns404() {
+        String challengeId = "nonExisting_challengeId";
+        String userId = "existing_userId";
+        String authHeader = "validAuthHeader";
+
+        String errorMessage = "ErrorMessage";
+
+        when(challengeService.removeChallengeFromFavorites(challengeId, userId)).thenReturn(Mono.error(new ChallengeNotFoundReturn404Exception(errorMessage)));
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenReturn(userId);
+
+        webTestClient.delete()
+                .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
+                .header("Authorization", authHeader)
+                .exchange()
+                .expectStatus().isNotFound()
+                .expectBody(MessageDto.class)
+                .value(messageDto -> Assertions.assertEquals(errorMessage, messageDto.getMessage()));
+
+        verify(challengeService, times(1)).removeChallengeFromFavorites(challengeId, userId);
+    }
+
+    @Test
+    void removeChallengeFromFavorite_InternalServerError_Returns500() {
+        String challengeId = "Existing_challengeId";
+        String userId = "existing_userId";
+        String authHeader = "validAuthHeader";
+
+        String errorMessage = "ErrorMessage";
+
+        when(challengeService.removeChallengeFromFavorites(challengeId, userId)).thenReturn(Mono.error(new InternalServerErrorException(errorMessage)));
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenReturn(userId);
+
+        webTestClient.delete()
+                .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
+                .header("Authorization", authHeader)
+                .exchange()
+                .expectStatus()
+                .isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
+                .expectBody(MessageDto.class)
+                .value(messageDto -> Assertions.assertEquals(errorMessage, messageDto.getMessage()));
+
+        verify(challengeService, times(1)).removeChallengeFromFavorites(challengeId, userId);
+    }
+
+    @Test
+    void removeChallengeFromFavorite_InvalidHeader_Returns400() {
+        String challengeId = "Existing_challengeId";
+        String authHeader = "BadHeader";
+        String errorMessage = "Error message";
+
+        when(jwtService.getUserUuIdFromAuthenticationHeader(authHeader)).thenThrow(new JwtException(errorMessage));
+
+        webTestClient.delete()
+                .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
+                .header("Authorization", authHeader)
+                .exchange()
+                .expectStatus()
+                .isEqualTo(HttpStatus.BAD_REQUEST)
+                .expectBody(MessageDto.class)
+                .value(messageDto -> Assertions.assertEquals(errorMessage, messageDto.getMessage()));
+
+        verify(challengeService, times(0)).removeChallengeFromFavorites(anyString(), anyString());
+    }
+
+    @Test
+    void removeChallengeFromFavorite_MissingHeader_Returns400() {
+        String challengeId = "Existing_challengeId";
+        String errorMessage = "ErrorMessage";
+
+        when(jwtService.getUserUuIdFromAuthenticationHeader(null)).thenThrow(new JwtException(errorMessage));
+
+        webTestClient.delete()
+                .uri("/itachallenge/api/v1/challenge/challenges/" + challengeId + "/favorites")
+                .exchange()
+                .expectStatus()
+                .isEqualTo(HttpStatus.BAD_REQUEST)
+                .expectBody(MessageDto.class)
+                .value(messageDto -> Assertions.assertEquals(errorMessage, messageDto.getMessage()));
+
+        verify(challengeService, times(0)).removeChallengeFromFavorites(anyString(), anyString());
+
+    }
+
+    @Test
+    @DisplayName("GET /challenges must return the timesFavorite field in the JSON")
+    void getChallenges_IncludesTimesFavorite() {
+        ChallengeDto challenge = ChallengeDto.builder()
+                .challengeId(UUID.randomUUID())
+                .title("Repte amb cor")
+                .level("Hard")
+                .creationDate("2025-03-26")
+                .detail(new DetailDocument("detall"))
+                .languages(Set.of())
+                .solutions(List.of())
+                .topic(Topic.DEBUGGING)
+                .timesFavorite(5)
+                .build();
+
+        ChallengeDto[] challengeArray = new ChallengeDto[] { challenge };
+        GenericResultDto<ChallengeDto> resultDto = new GenericResultDto<>(0, 10, 1, challengeArray);
+
+
+        when(challengeService.getAllChallenges(anyInt(), anyInt()))
+                .thenReturn(Mono.just(resultDto));
+
+        webTestClient.get()
+                .uri("/itachallenge/api/v1/challenge/challenges")
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody()
+                .jsonPath("$.results[0].timesFavorite").isEqualTo(5);
     }
 
 }