+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/document/TagTest.java
@@ -2,7 +2,11 @@ package com.itachallenge.challenge.document;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.UUID;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 public class TagTest {
 
@@ -11,7 +15,6 @@ public class TagTest {
         String tagNameTest = "POO";
         TagDocument tag = new TagDocument(null, tagNameTest, null);
         assertEquals(tagNameTest, tag.getTagName());
-
     }
 
     @Test
@@ -19,7 +22,6 @@ public class TagTest {
         String tagDescriptionTest = "Programació orientada a objectes";
         TagDocument tag = new TagDocument(null, "POO", tagDescriptionTest);
         assertEquals(tagDescriptionTest, tag.getTagDescription());
-
     }
 
     @Test
@@ -29,7 +31,6 @@ public class TagTest {
         TagDocument tag = new TagDocument(null, firstTagName, null);
         tag.setTagName(tagNameTest);
         assertEquals(tagNameTest, tag.getTagName());
-
     }
 
     @Test
@@ -39,6 +40,19 @@ public class TagTest {
         TagDocument tag = new TagDocument(null, "POO", firstTagDescription);
         tag.setTagDescription(tagDescriptionTest);
         assertEquals(tagDescriptionTest, tag.getTagDescription());
-
     }
+
+    @Test
+    void fullConstructorTest() {
+        UUID id = UUID.randomUUID();
+        String name = "POO";
+        String desc = "Programació orientada a objectes";
+
+        TagDocument tag = new TagDocument(id, name, desc);
+
+        assertEquals(id, tag.getIdTag());
+        assertEquals(name, tag.getTagName());
+        assertEquals(desc, tag.getTagDescription());
+    }
+    
 }