+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/integration/ChallengeIntegrationTest.java
@@ -16,6 +16,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.web.reactive.server.WebTestClient;
+import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.MongoDBContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -26,6 +27,7 @@ import java.time.LocalDateTime;
 import java.util.*;
 
 import static org.hamcrest.Matchers.equalTo;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpStatus.OK;
 
@@ -65,7 +67,7 @@ class ChallengeIntegrationTest {
         UUID uuidLang2 = UUID.fromString("409c9fe8-74de-4db3-81a1-a55280cf92ef");
         UUID[] idsLanguages = new UUID[]{uuidLang1, uuidLang2};
         String[] languageNames = new String[]{"name1", "name2"};
-        List<TagDocument> tags = List.of(new TagDocument());
+        List<UUID> tags = List.of(UUID.randomUUID());
         String languageImage = "https://image-default.com/default.png";
         LanguageDocument language1 = getLanguageMocked(idsLanguages[0], languageNames[0], languageImage);
         LanguageDocument language2 = getLanguageMocked(idsLanguages[1], languageNames[1], languageImage);
@@ -158,4 +160,6 @@ class ChallengeIntegrationTest {
                 .contains(new ChallengeDto[]{})
                 .hasSize(1);
     }
+
+
 }