+++ itachallenge-user/src/test/java/com/itachallenge/user/dto/UserSolutionRequestDtoTest.java
@@ -1,0 +1,98 @@
+package com.itachallenge.user.dto;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.UUID;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+class UserSolutionRequestDtoTest {
+
+    private UserSolutionRequestDto userSolutionRequestDto;
+    String userID = UUID.randomUUID().toString();
+    String challengeId = UUID.randomUUID().toString();
+    String languageId = UUID.randomUUID().toString();
+    String status = "status";
+    String solutionText = "This is my solution";
+    UserSolutionRequestDto solutionDto = new UserSolutionRequestDto();
+
+    @BeforeEach
+    public void setUp() {
+        userSolutionRequestDto = UserSolutionRequestDto.builder()
+                .userId("validUserId")
+                .challengeId("validChallengeId")
+                .languageId("validLanguageId")
+                .status("validChallengeStatus")
+                .solutionText("Valid solution text")
+                .build();
+    }
+
+    @Test
+    void testLombokGeneratedMethods() {
+        UserSolutionRequestDto dto1 = UserSolutionRequestDto.builder().build();
+
+        assertThat(dto1).isNotNull();
+        assertThat(dto1.toString()).isNotEmpty();
+        assertThat(dto1.getClass()).isEqualTo(UserSolutionRequestDto.class);
+    }
+
+    @Test
+    void getterUserSolutionDto_test() {
+        assertNotNull(userSolutionRequestDto);
+        assertEquals("validUserId", userSolutionRequestDto.getUserId());
+        assertEquals("validChallengeId", userSolutionRequestDto.getChallengeId());
+        assertEquals("validLanguageId", userSolutionRequestDto.getLanguageId());
+        assertEquals("Valid solution text", userSolutionRequestDto.getSolutionText());
+        assertEquals("validChallengeStatus", userSolutionRequestDto.getStatus());
+    }
+
+    @Test
+    void noArgsConstructor_GetterAndSetter_UserSolutionDto_test(){
+        solutionDto.setUserId(userID);
+        solutionDto.setChallengeId(challengeId);
+        solutionDto.setLanguageId(languageId);
+        solutionDto.setStatus(status);
+        solutionDto.setSolutionText(solutionText);
+
+        assertThat(solutionDto.getUserId()).isEqualTo(userID);
+        assertThat(solutionDto.getChallengeId()).isEqualTo(challengeId);
+        assertThat(solutionDto.getLanguageId()).isEqualTo(languageId);
+        assertThat(solutionDto.getStatus()).isEqualTo(status);
+        assertThat(solutionDto.getSolutionText()).isEqualTo(solutionText);
+    }
+
+    @Test
+    void jsonSerialization_test() throws Exception{
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(userSolutionRequestDto);
+        assertTrue(json.contains("\"uuid_user\":\"validUserId\""));
+        assertTrue(json.contains("\"uuid_language\":\"validLanguageId\""));
+        assertTrue(json.contains("\"uuid_challenge\":\"validChallengeId\""));
+        assertTrue(json.contains("\"solution_text\":\"Valid solution text\""));
+        assertTrue(json.contains("\"status\":\"validChallengeStatus\""));
+    }
+    @Test
+    void requiredArgsConstructor_userSolutionScoreDto_test(){
+        UserSolutionRequestDto userSolutionRequestDto1 = new UserSolutionRequestDto(
+                userID, challengeId, languageId, status, solutionText);
+        assertThat(userSolutionRequestDto1.getUserId()).isEqualTo(userID);
+        assertThat(userSolutionRequestDto1.getChallengeId()).isEqualTo(challengeId);
+        assertThat(userSolutionRequestDto1.getLanguageId()).isEqualTo(languageId);
+        assertThat(userSolutionRequestDto1.getStatus()).isEqualTo(status);
+        assertThat(userSolutionRequestDto1.getSolutionText()).isEqualTo(solutionText);
+    }
+
+    @Test
+    void testInvalidUserId() {
+        userSolutionRequestDto.setUserId("invalidUserId");
+        assertEquals("invalidUserId", userSolutionRequestDto.getUserId());
+    }
+
+    @Test
+    void testInvalidSolutionText() {
+        userSolutionRequestDto.setSolutionText("");
+        assertEquals("", userSolutionRequestDto.getSolutionText());
+    }
+}